Extract payment service
