+++ src/main/java/uk/gov/hmcts/divorce/citizen/event/CitizenSubmitApplication.java
@@ -12,8 +12,8 @@ import uk.gov.hmcts.ccd.sdk.type.OrderSummary;
 import uk.gov.hmcts.divorce.common.model.CaseData;
 import uk.gov.hmcts.divorce.common.model.State;
 import uk.gov.hmcts.divorce.common.model.UserRole;
+import uk.gov.hmcts.divorce.payment.PaymentService;
 import uk.gov.hmcts.divorce.payment.model.Payment;
-import uk.gov.hmcts.divorce.solicitor.service.SolicitorSubmitApplicationService;
 
 import java.util.List;
 import java.util.UUID;
@@ -34,7 +34,7 @@ public class CitizenSubmitApplication implements CCDConfig<CaseData, State, User
     public static final String CITIZEN_SUBMIT = "citizen-submit-application";
 
     @Autowired
-    private SolicitorSubmitApplicationService solicitorSubmitApplicationService;
+    private PaymentService paymentService;
 
     @Override
     public void configure(ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -55,7 +55,7 @@ public class CitizenSubmitApplication implements CCDConfig<CaseData, State, User
         log.info("Submit application about to start callback invoked");
 
         CaseData caseDataCopy = details.getData().toBuilder().build();
-        OrderSummary orderSummary = solicitorSubmitApplicationService.getOrderSummary();
+        OrderSummary orderSummary = paymentService.getOrderSummary();
         caseDataCopy.setApplicationFeeOrderSummary(orderSummary);
 
         ListValue<Payment> paymentListValue = createPendingPayment(orderSummary.getPaymentTotal());
