+++ src/main/java/uk/gov/hmcts/divorce/payment/PaymentService.java
@@ -1,0 +1,58 @@
+package uk.gov.hmcts.divorce.payment;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import uk.gov.hmcts.ccd.sdk.type.Fee;
+import uk.gov.hmcts.ccd.sdk.type.ListValue;
+import uk.gov.hmcts.ccd.sdk.type.OrderSummary;
+import uk.gov.hmcts.divorce.payment.model.FeeResponse;
+import uk.gov.hmcts.divorce.payment.model.Payment;
+
+import static java.util.Collections.singletonList;
+import static uk.gov.hmcts.ccd.sdk.type.Fee.getValueInPence;
+import static uk.gov.hmcts.divorce.payment.model.PaymentStatus.SUCCESS;
+
+@Service
+public class PaymentService {
+
+    private static final String DEFAULT_CHANNEL = "default";
+    private static final String ISSUE_EVENT = "issue";
+    private static final String FAMILY = "family";
+    private static final String FAMILY_COURT = "family court";
+    private static final String DIVORCE = "divorce";
+
+    @Autowired
+    private FeesAndPaymentsClient feesAndPaymentsClient;
+
+    public OrderSummary getOrderSummary() {
+        final var feeResponse = feesAndPaymentsClient.getApplicationIssueFee(
+            DEFAULT_CHANNEL,
+            ISSUE_EVENT,
+            FAMILY,
+            FAMILY_COURT,
+            DIVORCE,
+            null
+        );
+
+        return OrderSummary
+            .builder()
+            .fees(singletonList(getFee(feeResponse)))
+            .paymentTotal(getValueInPence(feeResponse.getAmount()))
+            .build();
+    }
+
+    private ListValue<Fee> getFee(final FeeResponse feeResponse) {
+        return ListValue
+            .<Fee>builder()
+            .value(
+                Fee
+                    .builder()
+                    .amount(getValueInPence(feeResponse.getAmount()))
+                    .code(feeResponse.getFeeCode())
+                    .description(feeResponse.getDescription())
+                    .version(String.valueOf(feeResponse.getVersion()))
+                    .build()
+            )
+            .build();
+    }
+}
