+++ src/main/java/uk/gov/hmcts/divorce/solicitor/service/SolicitorSubmitApplicationService.java
@@ -35,15 +35,6 @@ import static uk.gov.hmcts.divorce.payment.model.PaymentStatus.SUCCESS;
 @Slf4j
 public class SolicitorSubmitApplicationService {
 
-    private static final String DEFAULT_CHANNEL = "default";
-    private static final String ISSUE_EVENT = "issue";
-    private static final String FAMILY = "family";
-    private static final String FAMILY_COURT = "family court";
-    private static final String DIVORCE = "divorce";
-
-    @Autowired
-    private FeesAndPaymentsClient feesAndPaymentsClient;
-
     @Autowired
     private CaseDataUpdaterChainFactory caseDataUpdaterChainFactory;
 
@@ -55,23 +46,6 @@ public class SolicitorSubmitApplicationService {
 
     @Autowired
     private Clock clock;
-
-    public OrderSummary getOrderSummary() {
-        final var feeResponse = feesAndPaymentsClient.getApplicationIssueFee(
-            DEFAULT_CHANNEL,
-            ISSUE_EVENT,
-            FAMILY,
-            FAMILY_COURT,
-            DIVORCE,
-            null
-        );
-
-        return OrderSummary
-            .builder()
-            .fees(singletonList(getFee(feeResponse)))
-            .paymentTotal(getValueInPence(feeResponse.getAmount()))
-            .build();
-    }
 
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(
         final CaseData caseData,
@@ -113,21 +87,6 @@ public class SolicitorSubmitApplicationService {
             .data(updatedCaseData)
             .state(state)
             .errors(submittedErrors)
-            .build();
-    }
-
-    private ListValue<Fee> getFee(final FeeResponse feeResponse) {
-        return ListValue
-            .<Fee>builder()
-            .value(
-                Fee
-                    .builder()
-                    .amount(getValueInPence(feeResponse.getAmount()))
-                    .code(feeResponse.getFeeCode())
-                    .description(feeResponse.getDescription())
-                    .version(String.valueOf(feeResponse.getVersion()))
-                    .build()
-            )
             .build();
     }
